+++ x/evm/handler.go
@@ -74,7 +74,10 @@ func handleMsgEthereumTx(ctx sdk.Context, k Keeper, msg types.MsgEthereumTx) (*s
 	k.Bloom.Or(k.Bloom, executionResult.Bloom)
 
 	// update transaction logs in KVStore
-	k.SetTransactionLogs(ctx, txHash, executionResult.Logs)
+	err = k.SetLogs(ctx, common.BytesToHash(txHash), executionResult.Logs)
+	if err != nil {
+		panic(err)
+	}
 
 	// log successful execution
 	k.Logger(ctx).Info(executionResult.Result.Log)
@@ -147,7 +150,10 @@ func handleMsgEthermint(ctx sdk.Context, k Keeper, msg types.MsgEthermint) (*sdk
 	k.Bloom.Or(k.Bloom, executionResult.Bloom)
 
 	// update transaction logs in KVStore
-	k.SetTransactionLogs(ctx, txHash, executionResult.Logs)
+	err = k.SetLogs(ctx, common.BytesToHash(txHash), executionResult.Logs)
+	if err != nil {
+		panic(err)
+	}
 
 	// log successful execution
 	k.Logger(ctx).Info(executionResult.Result.Log)