+++ x/evm/types/journal.go
@@ -204,11 +204,16 @@ func (ch refundChange) dirtied() *ethcmn.Address {
 }
 
 func (ch addLogChange) revert(s *CommitStateDB) {
-	logs := s.logs[ch.txhash]
+	logs, err := s.GetLogs(ch.txhash)
+	if err != nil {
+		// panic on unmarshal error
+		panic(err)
+	}
+
 	if len(logs) == 1 {
-		delete(s.logs, ch.txhash)
-	} else {
-		s.logs[ch.txhash] = logs[:len(logs)-1]
+		s.DeleteLogs(ch.txhash)
+	} else if err := s.SetLogs(ch.txhash, logs[:len(logs)-1]); err != nil {
+		panic(err)
 	}
 
 	s.logSize--